food-delivery-chatbot/requirements.txt
# Core dependencies
openai>=1.12.0
langchain>=0.1.0
langchain-openai>=0.0.5
python-dotenv>=1.0.0

# Web interface
streamlit>=1.31.0

# Data handling
pandas>=2.0.0

# Optional: For voice support
# speechrecognition>=3.10.0
# pyttsx3>=2.90
